% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllIdentical.R
\name{AllIdentical}
\alias{AllIdentical}
\title{AllIdentical}
\usage{
AllIdentical(...)
}
\arguments{
\item{...}{a list of object of arbitrary length}
}
\value{
logical. TRUE if all objects are identical
}
\description{
A faster way to determine if all objects in a list are identical
}
