% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_exchange.R
\name{author_exchange}
\alias{author_exchange}
\title{This function exchanges full author list formated with First and Middle initials
and full family name. This allows the creation of bib entries with full author
lists}
\usage{
author_exchange(x, y)
}
\arguments{
\item{x}{bib entry created from RefManageR::ReadGS() (i.e., list) with incomplete
author list}

\item{y}{personList created from convert_complete_author(). Same length as x}
}
\description{
This function exchanges full author list formated with First and Middle initials
and full family name. This allows the creation of bib entries with full author
lists
}
\examples{
author_exchange()
#mapply(author_exchange, x = jrj.bib, y = full_authors_list)
}
