% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_complete_authors.R
\name{convert_complete_author}
\alias{convert_complete_author}
\title{This function converts full author lists to person objects.}
\usage{
convert_complete_author(complete_coauthor_string)
}
\arguments{
\item{complete_coauthor_string}{A character string of the full author list
separated by ,s of the form e.g., James Robert Junker, John Jacob Jingle Heimer-Schimdt, etc.}
}
\description{
It can be used to import full author person objects into bib objects
returned from RefManageR::ReadGS. This further compresses the full author
names into just the first character for given and middle names.
}
