% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebplot.R
\name{create_xy}
\alias{create_xy}
\title{create_xy}
\usage{
create_xy(po)
}
\arguments{
\item{po}{the number of nodes in a food web passed from foodwebplot.R}
}
\value{
returns two columns of x and y coordinates for the number of nodes in a food web
}
\description{
This function is a support function for foodwebplot to create regularly spaced circular coordinates for the size of a given food web
}
