% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomIsoline.R
\docType{data}
\name{GeomIsoline}
\alias{GeomIsoline}
\title{geom_isolineNP
This geom adds NP stoichiometric isolines to plot background for NP pools}
\format{
An object of class \code{GeomIsoline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomIsoline
}
\description{
geom_isolineNP
This geom adds NP stoichiometric isolines to plot background for NP pools
}
\keyword{datasets}
