% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complete_coauthors.R
\name{get_complete_coauthors}
\alias{get_complete_coauthors}
\title{This function scrapes the full coauthor list for a publication from
Google Scholar. It is a modified version from the scholar package that
includes a sleep option to allow a time between http requests.}
\usage{
get_complete_coauthors(id, pubid, sleep = 3, ...)
}
\arguments{
\item{id}{A Google scholar id from a public id page}

\item{pubid}{A publication id can be returned from get_publications()}

\item{sleep}{A number in seconds to wait in between requests passes to Sys.sleep()}

\item{...}{further argments to pass such as agent identification, encoding, etc.}
}
\value{
A character string with a full author list of full author names
}
\description{
This function scrapes the full coauthor list for a publication from
Google Scholar. It is a modified version from the scholar package that
includes a sleep option to allow a time between http requests.
}
