% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{\%ni\%}
\alias{\%ni\%}
\title{\%ni\%}
\usage{
x \%ni\% table
}
\arguments{
\item{x}{vector or NULL: the values to exclude. Long vectors are supported.}

\item{table}{vector or NULL: the values to be excluded against. Long vectors are not supported.
#' @returns A logical vector, indicating if a match was not located for each element of x: thus the values are TRUE or FALSE and never NA.}
}
\description{
This function is the opposite of \code{\%in\%} in that it finds items not in a vector
}
