% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\name{\%||\%}
\alias{\%||\%}
\title{\%\|\|\%}
\usage{
a \%||\% b
}
\arguments{
\item{...}{further arguments, including the variable/s by which to group_split and set names}
}
\description{
This function/Geom contains the code to create half-violin plots
}
\details{
geom_flat_violin
}
\examples{
geom_flat_violin()
### Example:
ggplot(diamonds, aes(cut, carat)) +
geom_flat_violin() +
coord_flip()
}
