% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_group_split.R
\name{named_group_split}
\alias{named_group_split}
\title{named_group_split}
\usage{
named_group_split(.tbl, ...)
}
\arguments{
\item{.tbl}{A tibble or tibble-convertable object}

\item{...}{further arguments, including the variable/s by which to group_split and set names}
}
\description{
This function returns a named list split on a grouping variable. It is based on the comments here: https://github.com/tidyverse/dplyr/issues/4223#issuecomment-469269857
}
\examples{
named_group_split()
}
