% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdown_addins.R
\name{trackdown_download}
\alias{trackdown_download}
\title{trackdown_download}
\usage{
trackdown_download(dl_filePattern = "download_trackdown.R", force = FALSE, ...)
}
\arguments{
\item{dl_filePattern}{the file pattern for the script used to automate the download}

\item{force}{logical. should the script force the download if multiple scripts are located}

\item{...}{additional arguments passed to function}
}
\value{
nothing
}
\description{
This function automates a trackdown download from google drive based on the script of \code{download_trackdown.R}
}
